Restructured the directory & updated the test folder
