+++ lib/screens/homepage.dart
@@ -1,0 +1,176 @@
+import 'package:flutter/material.dart';
+import 'loginpage.dart';
+
+class HomePage extends StatelessWidget {
+  final bool isUserRegistered = false;
+
+  @override
+  Widget build(BuildContext context) {
+    double screenWidth = MediaQuery.of(context).size.width;
+
+    return Scaffold(
+      appBar: PreferredSize(
+        preferredSize: Size.fromHeight(91),
+        child: Container(
+          color: const Color.fromARGB(255, 255, 255, 255),
+          child: Stack(
+            children: [
+              Positioned(
+                left: screenWidth * (33 / 390),
+                top: 35,
+                child: Text(
+                  'Shopifty',
+                  style: TextStyle(
+                    fontSize: 24,
+                    fontWeight: FontWeight.bold,
+                    fontFamily: 'Inria Sans',
+                    color: Colors.black,
+                  ),
+                ),
+              ),
+              Positioned(
+                left: screenWidth * (330.06 / 390),
+                top: 39.37,
+                child: IconButton(
+                  icon: Icon(
+                    Icons.shopping_cart,
+                    color: const Color.fromARGB(255, 0, 0, 0),
+                  ),
+                  iconSize: 24.3,
+                  onPressed: () {
+                    print('To Carts Page!');
+                  },
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 20.0),
+        child: Row(
+          children: [
+            Expanded(
+              child: Container(
+                height: 48,
+                padding: EdgeInsets.symmetric(horizontal: 12),
+                decoration: BoxDecoration(
+                  color: Colors.grey[200],
+                  borderRadius: BorderRadius.circular(8),
+                ),
+                child: Row(
+                  children: [
+                    Icon(
+                      Icons.search,
+                      color: const Color.fromRGBO(158, 158, 158, 1),
+                    ),
+                    SizedBox(width: 8),
+                    Expanded(
+                      child: TextField(
+                        decoration: InputDecoration(
+                          hintText: 'Search Product...',
+                          hintStyle: TextStyle(color: Colors.grey),
+                          border: InputBorder.none,
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+            SizedBox(width: 12),
+            Container(
+              height: 48,
+              child: ElevatedButton(
+                style: ElevatedButton.styleFrom(
+                  backgroundColor: Colors.black,
+                  foregroundColor: Colors.white,
+                  padding: EdgeInsets.symmetric(horizontal: 20),
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(8),
+                  ),
+                ),
+                onPressed: () {
+                  print('Search Button!');
+                },
+                child: Icon(Icons.search, size: 24),
+              ),
+            ),
+          ],
+        ),
+      ),
+      bottomNavigationBar: BottomNavigationBar(
+        backgroundColor: Colors.white,
+        selectedItemColor: Colors.black,
+        unselectedItemColor: Colors.grey,
+        onTap: (index) {
+          switch (index) {
+            case 0:
+              Navigator.push(
+                context,
+                MaterialPageRoute(builder: (_) => HomePlaceholderPage()),
+              );
+              break;
+            case 1:
+              Navigator.push(
+                context,
+                MaterialPageRoute(builder: (_) => BookmarkPage()),
+              );
+              break;
+            case 2:
+              if (isUserRegistered) {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (_) => ProfilePage()),
+                );
+              } else {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (_) => LoginPage()),
+                );
+              }
+              break;
+          }
+        },
+        items: const [
+          BottomNavigationBarItem(icon: Icon(Icons.home), label: 'Home'),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.bookmark_border),
+            label: 'Bookmarks',
+          ),
+          BottomNavigationBarItem(icon: Icon(Icons.person), label: 'Profile'),
+        ],
+      ),
+    );
+  }
+}
+
+class HomePlaceholderPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text('Home Placeholder')),
+      body: Center(child: Text('This is the Home page')),
+    );
+  }
+}
+
+class BookmarkPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text('Bookmarks')),
+      body: Center(child: Text('This is the Bookmark page')),
+    );
+  }
+}
+
+class ProfilePage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text('Your Profile')),
+      body: Center(child: Text('This is the Profile page')),
+    );
+  }
+}
