--- lib/screens/splash.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-
-class SplashScreen extends StatelessWidget {
-  const SplashScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-
-    );
-  }
-}
