+++ lib/screens/splashscreen.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+import 'homepage.dart'; // Make sure this matches your import path
+
+class SplashScreen extends StatefulWidget {
+  @override
+  _SplashScreenState createState() => _SplashScreenState();
+}
+
+class _SplashScreenState extends State<SplashScreen> {
+  @override
+  void initState() {
+    super.initState();
+    // Navigate to HomePage after 2 seconds
+    Future.delayed(Duration(seconds: 2), () {
+      Navigator.pushReplacement(
+        context,
+        MaterialPageRoute(builder: (_) => HomePage()),
+      );
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white, // Customize background
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(
+              Icons.shopping_bag,
+              size: 80,
+              color: Colors.black,
+            ), // Logo/Icon
+            SizedBox(height: 20),
+            Text(
+              'Shopifty',
+              style: TextStyle(
+                fontSize: 28,
+                fontWeight: FontWeight.bold,
+                fontFamily: 'Inria Sans',
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
